pyproject.toml
[project]
name = "nonebot-plugin-auto-bot-selector"
version = "0.3.0"
description = "自动Bot选择器，消息推送好帮手"
authors = [{ name = "Well404", email = "well_404@outlook.com" }]
dependencies = ["nonebot2>=2.2.1", "typing-extensions>=4.11.0"]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }
[project.optional-dependencies]
dev = [
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot-adapter-qq>=1.4.3",
    "nonebot-adapter-red>=0.9.0",
    "nonebot-adapter-satori>=0.10.4",
    "nonebot-adapter-dodo>=0.2.1",
    "nonebot-adapter-kaiheila>=0.3.3",
    "nonebot-plugin-alconna>=0.42.3",
    "nonebot-plugin-send-anything-anywhere>=0.6.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pyright]
pythonVersion = "3.8"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true, USE_SAA = false}
